requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
alembic==1.13.0
qrcode==7.4.2
pillow==10.1.0
requests==2.31.0
aiohttp==3.9.1
